+++ packages/event/src/BailEvent.ts
@@ -1,7 +1,9 @@
 import BaseEvent from './BaseEvent';
-import { Scope } from './types';
 
-export default class BailEvent<Args extends unknown[]> extends BaseEvent<Args, boolean | void> {
+export default class BailEvent<
+  Args extends unknown[],
+  Scope extends string = string
+> extends BaseEvent<boolean | void, Args, Scope> {
   /**
    * Synchronously execute listeners with the defined arguments.
    * If a listener returns `false`, the loop with be aborted early,